Refactor x/distribution according to ADR 031 (#7524)
* Refactor x/distribution according to ADR 31

* lint

* removed unused

* Apply suggestions from code review
