+++ proto/cosmos/distribution/v1beta1/tx.proto
@@ -7,6 +7,25 @@ option (gogoproto.equal_all) = true;
 import "gogoproto/gogo.proto";
 import "cosmos/base/v1beta1/coin.proto";
 
+// Msg defines the distribution Msg service.
+service Msg {
+  // SetWithdrawAddress defines a method to change the withdraw address 
+  // for a delegator (or validator self-delegation).
+  rpc SetWithdrawAddress(MsgSetWithdrawAddress) returns (MsgSetWithdrawAddressResponse);
+
+  // WithdrawDelegatorReward defines a method to withdraw rewards of delegator
+  // from a single validator.
+  rpc WithdrawDelegatorReward(MsgWithdrawDelegatorReward) returns (MsgWithdrawDelegatorRewardResponse);
+
+  // WithdrawValidatorCommission defines a method to withdraw the 
+  // full commission to the validator address.
+  rpc WithdrawValidatorCommission(MsgWithdrawValidatorCommission) returns (MsgWithdrawValidatorCommissionResponse);
+
+  // FundCommunityPool defines a method to allow an account to directly
+  // fund the community pool.
+  rpc FundCommunityPool(MsgFundCommunityPool) returns (MsgFundCommunityPoolResponse);
+}
+
 // MsgSetWithdrawAddress sets the withdraw address for
 // a delegator (or validator self-delegation).
 message MsgSetWithdrawAddress {
@@ -16,6 +35,9 @@ message MsgSetWithdrawAddress {
   string delegator_address = 1 [(gogoproto.moretags) = "yaml:\"delegator_address\""];
   string withdraw_address  = 2 [(gogoproto.moretags) = "yaml:\"withdraw_address\""];
 }
+
+// MsgSetWithdrawAddressResponse defines the Msg/SetWithdrawAddress response type.
+message MsgSetWithdrawAddressResponse { }
 
 // MsgWithdrawDelegatorReward represents delegation withdrawal to a delegator
 // from a single validator.
@@ -27,6 +49,9 @@ message MsgWithdrawDelegatorReward {
   string validator_address = 2 [(gogoproto.moretags) = "yaml:\"validator_address\""];
 }
 
+// MsgWithdrawDelegatorRewardResponse defines the Msg/WithdrawDelegatorReward response type.
+message MsgWithdrawDelegatorRewardResponse { }
+
 // MsgWithdrawValidatorCommission withdraws the full commission to the validator
 // address.
 message MsgWithdrawValidatorCommission {
@@ -35,6 +60,9 @@ message MsgWithdrawValidatorCommission {
 
   string validator_address = 1 [(gogoproto.moretags) = "yaml:\"validator_address\""];
 }
+
+// MsgWithdrawValidatorCommissionResponse defines the Msg/WithdrawValidatorCommission response type.
+message MsgWithdrawValidatorCommissionResponse { }
 
 // MsgFundCommunityPool allows an account to directly
 // fund the community pool.
@@ -46,3 +74,6 @@ message MsgFundCommunityPool {
       [(gogoproto.nullable) = false, (gogoproto.castrepeated) = "github.com/cosmos/cosmos-sdk/types.Coins"];
   string depositor = 2;
 }
+
+// MsgFundCommunityPoolResponse defines the Msg/FundCommunityPool response type.
+message MsgFundCommunityPoolResponse { }
